chore: incorrect stargaze gas, refactor config
