+++ src/config/chain/index.ts
@@ -2,7 +2,7 @@ import { NEUTRON_CHAIN_CONFIGS } from "./neutron";
 import { OSMOSIS_CHAIN_CONFIGS } from "./osmosis";
 import { SEI_CHAIN_CONFIGS } from "./sei";
 import { STARGAZE_CHAIN_CONFIGS } from "./stargaze";
-import type { ChainConfigs, ProjectConstants } from "./types";
+import type { ChainConfigs } from "./types";
 
 export * from "./types";
 export * from "./default";
@@ -12,12 +12,4 @@ export const CHAIN_CONFIGS: ChainConfigs = {
   ...SEI_CHAIN_CONFIGS,
   ...NEUTRON_CHAIN_CONFIGS,
   ...STARGAZE_CHAIN_CONFIGS,
-};
-
-export const PROJECT_CONSTANTS: ProjectConstants = {
-  maxListNameLength: 50,
-  maxContractNameLength: 50,
-  maxContractDescriptionLength: 250,
-  maxCodeNameLength: 50,
-  maxProposalTitleLength: 255,
 };
