+++ src/config/project/index.ts
@@ -1,0 +1,11 @@
+import type { ProjectConstants } from "./types";
+
+export * from "./types";
+
+export const PROJECT_CONSTANTS: ProjectConstants = {
+  maxListNameLength: 50,
+  maxContractNameLength: 50,
+  maxContractDescriptionLength: 250,
+  maxCodeNameLength: 50,
+  maxProposalTitleLength: 255,
+};
