+++ src/config/project/types.ts
@@ -1,0 +1,10 @@
+export interface ProjectConstants {
+  // wasm
+  maxListNameLength: number;
+  maxContractNameLength: number;
+  maxContractDescriptionLength: number;
+  maxCodeNameLength: number;
+
+  // proposal
+  maxProposalTitleLength: number;
+}
